Divide code into modules
