+++ src/modules/localStorage.js
@@ -1,6 +1,7 @@
 // -- Define local storage --
 
 export default class Storage {
+  // -- To define the LS array --
   static getTasks() {
     let tasks;
     if (localStorage.getItem('tasks') === null) {
@@ -10,13 +11,13 @@ export default class Storage {
     }
     return tasks;
   }
-
+// -- When pressed add task we call this functin to add it to LS --
   static addTask(task) {
     const tasks = Storage.getTasks();
     tasks.push(task);
     localStorage.setItem('tasks', JSON.stringify(tasks));
   }
-
+// -- When pressed remove task we call this functin to remove it from LS --
   static removeTask(id) {
     let tasks = JSON.parse(localStorage.getItem('tasks'));
     tasks = tasks.filter((e) => e.index.toString() !== id.toString());
@@ -26,7 +27,7 @@ export default class Storage {
     }
     localStorage.setItem('tasks', JSON.stringify(tasks));
   }
-
+// -- When pressed checkbox, we change the status complete of the task  to true in LS --
   static updateStatus(id) {
     const tasks = JSON.parse(localStorage.getItem('tasks'));
     for (let i = 0; i < tasks.length; i += 1) {
@@ -41,7 +42,6 @@ export default class Storage {
       }
     }
   }
-
   // -- Update Local Storage when mdoify task description --
   static editTask(newTitle, newDescription, newDate, id) {
     const tasks = JSON.parse(localStorage.getItem('tasks'));
