+++ src/style.css
@@ -76,7 +76,6 @@ body {
   top: 4px;
   left: 4px;
 }
-
 /* <!-- Sidebar menu ENDS --> */
 
 /* <!-- Main app  --> */
@@ -133,7 +132,7 @@ body {
   gap: 10px;
 }
 
-/* <!-- Modal add starts --> */
+/* <!-- Modal starts --> */
 .modal {
   position: fixed;
   z-index: 1;
@@ -283,8 +282,7 @@ textarea {
   font-size: 12px;
   font-family: 'Raleway', sans-serif;
 }
-
-/* <!-- Modal add ENDS --> */
+/* <!-- Modal  ENDS --> */
 
 /* <!-- Modal info starts --> */
 .modal-info-container {
@@ -320,7 +318,7 @@ textarea {
   background: linear-gradient(109.6deg, rgb(151, 213, 255) 11.2%, rgb(255, 155, 185) 91.1%);
   color: rgb(81, 81, 81);
 }
-
+/* <!-- Modal info ENDS --> */
 @media only screen and (min-width: 768px) {
   /* <!-- Sidebar menu STARTS --> */
   .close-menu-btn {
@@ -378,7 +376,6 @@ textarea {
     background-color: whitesmoke;
     border-radius: 10px;
   }
-
   /* <!-- Sidebar menu ENDS --> */
 
   /* <!-- Main app--> */
